+++ src/python/autotransform/command/script.py
@@ -82,7 +82,7 @@ class ScriptCommand(Command):
             if self.run_on_changes:
                 current_schema = autotransform.schema.current
                 assert current_schema is not None
-                repo = current_schema.get_repo()
+                repo = current_schema.repo
                 assert repo is not None
                 items: Sequence[Item] = [
                     FileItem(key=file) for file in repo.get_changed_files(batch)
@@ -182,7 +182,7 @@ class ScriptCommand(Command):
         if self.run_on_changes:
             current_schema = autotransform.schema.current
             assert current_schema is not None
-            repo = current_schema.get_repo()
+            repo = current_schema.repo
             assert repo is not None
             items: Sequence[Item] = [FileItem(key=file) for file in repo.get_changed_files(batch)]
         else: