+++ src/python/autotransform/step/condition/schema.py
@@ -49,7 +49,7 @@ class SchemaNameCondition(Condition):
             ComparisonType.NOT_EQUAL,
         ], "SchemaNameCondition may only use equal or not_equal comparison"
         return compare(
-            change.get_schema().get_config().schema_name,
+            change.get_schema().config.schema_name,
             self.schema_name,
             self.comparison,
         )